Move obtaining Java class for a field into the `Field` class
